Travis script improvements
A few improvements that should improve the travis script robustness and improve
execution time.

- Use `travis_wait` when running R CMD check. The OS X builds are often
failing due to a timeout. Using `travis_wait` will work around this issue
by generating output periodically, so the builds should complete.

- Use `sudo: false` and precise builds.

This will enable use of the container based infrastructure, which should
result in faster build times. Use of precise rather than trusty means
the dependencies will be earlier versions, but this may also ensure the
package works with more systems. Using `--user` for the pip calls should
also remove the warnings about invalid permissions
(https://travis-ci.org/rstudio/tensorflow/jobs/186765420#L749)

- Cache the pip dependencies as well as R dependencies.
Should speed up builds a decent amount.

- Rewrite the PATH splitting code

The current implementation uses 4 different processes and needs to use
workarounds to handle newlines, I think it is simpler to handle the path
properly in a scripting language. This commit uses perl, however if R or
python would be preferred equivalent implementations are.

    export PATH=$(Rscript -e 'cat(sep = ":", grep(invert = T, value = T, "/opt/python/", strsplit(Sys.getenv("PATH"), ":")[[1]]))')

    export PATH=$(python -c 'import os;import re;print ":".join(filter(lambda x:not re.search(r"/opt/python/", x), os.getenv("PATH").split(":")))')

- Fail the travis_install.sh script if any commands fail
Right now the travis build continues even when there are failures installing
tensorflow. Instead we set `set -e` which will fail the script, and
consequently the travis build if any of the commands fail.

- Update tenserflow URLs
The current URLs are returning a 404, this updates the URLs to the current version 0.12.1.

- Remove virtualenv calls
Virtualenv is not currently installed, so these are just throwing errors.