invoke R functions from Python with `tryCatch()`
Rcpp release 1.0.10 turned on use of `R_UnwindProtect()` by default.
Previous versions of Rcpp would eval R code with a `tryCatch()`,
silently catching errors instead of propagating them back up to the user
session. Reticulate was implicitly depending on this behavior in various
locations.

With this patch, reticulate explicitly places a `tryCatch()` on the
stack when Python is calling back into R, and then reticulate will raise
a python Exception if an R error was caught.

https://github.com/rstudio/reticulate/issues/1328