A different approach using lambdas
lookups of dunders for builtins don't always work from the class, and are slower due to going through __getattribute__ instead of an implicit lookup via the metaclass.
This approach brings us closer to the actual semantics in python, and also gives us a usually better python traceback (at the expense of a slightly uglier R traceback).