evaluate `call_r_func()` in user_env()
This causes `sys.parents()` (and by extension `rlang::trace_back()`) to return the parent of the call as something on the callstack, rather than `0`. This enables proper printing of the
error tracebacks as a leaf on a tree, rather than a top-level node. that's reset at each language boundary encountered on the stack.